SubLayout 컴포넌트 생성, 로그인/회원가입 중앙정렬, #root: width:100% 적용
