+++ src/index.css
@@ -8,6 +8,10 @@
   box-sizing: border-box;
 }
 
+#root {
+  width: 100%;
+}
+
 :root {
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
