+++ src/shared/Router.jsx
@@ -2,35 +2,20 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import ProtectedRoute from "../components/ProtectedRoute";
 import BookmarkTestHome from "../pages/bookmarkTest/bookmarkTestHome";
 import BookmarkTestMyPage from "../pages/bookmarkTest/bookmarkTestMyPage";
-import Layout from "../components/Layout";
 import Home from "../pages/Home";
 import Login from "../pages/join/Login";
 import Signup from "../pages/join/Signup";
 import MyPage from "../pages/MyPage";
 import { useState } from "react";
+import MainLayout from "../components/MainLayout";
+import SubLayout from "../components/SubLayout";
 
 const Router = () => {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Layout />}>
+        <Route path="/" element={<MainLayout />}>
           <Route index element={<Home />} />
-          <Route
-            path="/login"
-            element={
-              <ProtectedRoute>
-                <Login />
-              </ProtectedRoute>
-            }
-          />
-          <Route
-            path="/signup"
-            element={
-              <ProtectedRoute>
-                <Signup />
-              </ProtectedRoute>
-            }
-          />
           <Route
             path="/bookmarkhome"
             element={
@@ -48,14 +33,32 @@ const Router = () => {
             }
           />
         </Route>
-        <Route
-          path="/mypage"
-          element={
-            <ProtectedRoute authOnly={true}>
-              <MyPage />
-            </ProtectedRoute>
-          }
-        />
+        <Route path="/" element={<SubLayout />}>
+          <Route
+            path="/login"
+            element={
+              <ProtectedRoute>
+                <Login />
+              </ProtectedRoute>
+            }
+          />
+          <Route
+            path="/signup"
+            element={
+              <ProtectedRoute>
+                <Signup />
+              </ProtectedRoute>
+            }
+          />
+          <Route
+            path="/mypage"
+            element={
+              <ProtectedRoute authOnly={true}>
+                <MyPage />
+              </ProtectedRoute>
+            }
+          />
+        </Route>
       </Routes>
     </BrowserRouter>
   );
